+++ libs/journeys/ui/src/components/SearchBar/data.ts
@@ -51,3 +51,17 @@ export const languageRefinements: RefinementListItem[] = [
     highlighted: 'Chinese, Traditional'
   }
 ]
+
+interface JsonReturn {
+  country: string
+}
+
+interface FetchReturn {
+  json: () => Promise<JsonReturn>
+}
+
+export async function fetchCountryMock(): Promise<FetchReturn> {
+  return await Promise.resolve({
+    json: async () => await Promise.resolve({ country: 'US' })
+  })
+}
